+++ path/__init__.py
@@ -1,39 +1,26 @@
-#
-# Copyright (c) 2010 Mikhail Gusarov
-#
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-#
-# The above copyright notice and this permission notice shall be included in
-# all copies or substantial portions of the Software.
-#
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-# SOFTWARE.
-#
-
 """
-path.py - An object representing a path to a file or directory.
+Path Pie
 
-https://github.com/jaraco/path.py
+Implements ``path.Path`` - An object representing a
+path to a file or directory.
 
 Example::
 
     from path import Path
     d = Path('/home/guido/bin')
+
+    # Globbing
     for f in d.files('*.py'):
         f.chmod(0o755)
-"""
 
-from __future__ import unicode_literals
+    # Changing the working directory:
+    with Path("somewhere"):
+        # cwd in now `somewhere`
+        ...
+
+    # Concatenate paths with /
+    foo_txt = Path("bar") / "foo.txt"
+"""
 
 import sys
 import warnings
@@ -49,7 +36,6 @@ import operator
 import re
 import contextlib
 import io
-import distutils.dir_util
 import importlib
 import itertools
 
@@ -68,39 +54,11 @@ try:
 except ImportError:
     pass
 
-##############################################################################
-# Python 2/3 support
-PY3 = sys.version_info >= (3,)
-PY2 = not PY3
-
-string_types = str,
-text_type = str
-getcwdu = os.getcwd
+from . import matchers
+from .py37compat import best_realpath
 
 
-if PY2:
-    import __builtin__
-    string_types = __builtin__.basestring,
-    text_type = __builtin__.unicode
-    getcwdu = os.getcwdu
-    map = itertools.imap
-
-
-@contextlib.contextmanager
-def io_error_compat():
-    try:
-        yield
-    except IOError as io_err:
-        # On Python 2, io.open raises IOError; transform to OSError for
-        # future compatibility.
-        os_err = OSError(*io_err.args)
-        os_err.filename = getattr(io_err, 'filename', None)
-        raise os_err
-
-##############################################################################
-
-
-__all__ = ['Path', 'CaseInsensitivePattern']
+__all__ = ['Path', 'TempDir', 'CaseInsensitivePattern']
 
 
 LINESEPS = ['\r\n', '\r', '\n']
@@ -112,8 +70,15 @@ U_NL_END = re.compile(r'(?:{0})$'.format(U_NEWLINE.pattern))
 
 
 try:
-    import pkg_resources
-    __version__ = pkg_resources.require('path.py')[0].version
+    from importlib import metadata  # type: ignore
+except ImportError:
+    try:
+        import importlib_metadata as metadata  # type: ignore
+    except ImportError:
+        pass
+
+try:
+    __version__ = metadata.version('path')  # type: ignore
 except Exception:
     __version__ = 'unknown'
 
@@ -140,6 +105,7 @@ def simple_cache(func):
             return saved_results[module]
         saved_results[module] = func(cls, module)
         return saved_results[module]
+
     return wrapper
 
 
@@ -153,17 +119,19 @@ class multimethod(object):
     Acts like a classmethod when invoked from the class and like an
     instancemethod when invoked from the instance.
     """
+
     def __init__(self, func):
         self.func = func
 
     def __get__(self, instance, owner):
         return (
-            functools.partial(self.func, owner) if instance is None
+            functools.partial(self.func, owner)
+            if instance is None
             else functools.partial(self.func, owner, instance)
         )
 
 
-class Path(text_type):
+class Path(str):
     """
     Represents a filesystem path.
 
@@ -193,8 +161,6 @@ class Path(text_type):
     @simple_cache
     def using_module(cls, module):
         subclass_name = cls.__name__ + '_' + module.__name__
-        if PY2:
-            subclass_name = str(subclass_name)
         bases = (cls,)
         ns = {'module': module}
         return type(subclass_name, bases, ns)
@@ -220,7 +186,7 @@ class Path(text_type):
             return NotImplemented
 
     def __radd__(self, other):
-        if not isinstance(other, string_types):
+        if not isinstance(other, str):
             return NotImplemented
         return self._next_class(other.__add__(self))
 
@@ -269,7 +235,7 @@ class Path(text_type):
 
         .. seealso:: :func:`os.getcwdu`
         """
-        return cls(getcwdu())
+        return cls(os.getcwd())
 
     #
     # --- Operations on Path strings.
@@ -288,7 +254,8 @@ class Path(text_type):
 
     def realpath(self):
         """ .. seealso:: :func:`os.path.realpath` """
-        return self._next_class(self.module.realpath(self))
+        realpath = best_realpath(self.module)
+        return self._next_class(realpath(self))
 
     def expanduser(self):
         """ .. seealso:: :func:`os.path.expanduser` """
@@ -331,6 +298,25 @@ class Path(text_type):
         f, ext = self.module.splitext(self)
         return ext
 
+    def with_suffix(self, suffix):
+        """ Return a new path with the file suffix changed (or added, if none)
+
+        >>> Path('/home/guido/python.tar.gz').with_suffix(".foo")
+        Path('/home/guido/python.tar.foo')
+
+        >>> Path('python').with_suffix('.zip')
+        Path('python.zip')
+
+        >>> Path('filename.ext').with_suffix('zip')
+        Traceback (most recent call last):
+        ...
+        ValueError: Invalid suffix 'zip'
+        """
+        if not suffix.startswith('.'):
+            raise ValueError("Invalid suffix {suffix!r}".format(**locals()))
+
+        return self.stripext() + suffix
+
     @property
     def drive(self):
         """ The drive specifier, for example ``'C:'``.
@@ -341,7 +327,9 @@ class Path(text_type):
         return self._next_class(drive)
 
     parent = property(
-        dirname, None, None,
+        dirname,
+        None,
+        None,
         """ This path's parent directory, as a new Path object.
 
         For example,
@@ -349,17 +337,21 @@ class Path(text_type):
         Path('/usr/local/lib')``
 
         .. seealso:: :meth:`dirname`, :func:`os.path.dirname`
-        """)
+        """,
+    )
 
     name = property(
-        basename, None, None,
+        basename,
+        None,
+        None,
         """ The name of this file or directory without the full path.
 
         For example,
         ``Path('/usr/local/lib/libpython.so').name == 'libpython.so'``
 
         .. seealso:: :meth:`basename`, :func:`os.path.basename`
-        """)
+        """,
+    )
 
     def splitpath(self):
         """ p.splitpath() -> Return ``(p.parent, p.name)``.
@@ -501,7 +493,7 @@ class Path(text_type):
 
     # --- Listing, searching, walking, and matching
 
-    def listdir(self, pattern=None):
+    def listdir(self, match=None):
         """ D.listdir() -> List of items in this directory.
 
         Use :meth:`files` or :meth:`dirs` instead if you want a listing
@@ -509,46 +501,37 @@ class Path(text_type):
 
         The elements of the list are Path objects.
 
-        With the optional `pattern` argument, this only lists
-        items whose names match the given pattern.
+        With the optional `match` argument, a callable,
+        only return items whose names match the given pattern.
 
         .. seealso:: :meth:`files`, :meth:`dirs`
         """
-        if pattern is None:
-            pattern = '*'
-        return [
-            self / child
-            for child in os.listdir(self)
-            if self._next_class(child).fnmatch(pattern)
-        ]
+        match = matchers.load(match)
+        return list(filter(match, (self / child for child in os.listdir(self))))
 
-    def dirs(self, pattern=None):
+    def dirs(self, *args, **kwargs):
         """ D.dirs() -> List of this directory's subdirectories.
 
         The elements of the list are Path objects.
         This does not walk recursively into subdirectories
         (but see :meth:`walkdirs`).
 
-        With the optional `pattern` argument, this only lists
-        directories whose names match the given pattern.  For
-        example, ``d.dirs('build-*')``.
+        Accepts parameters to :meth:`listdir`.
         """
-        return [p for p in self.listdir(pattern) if p.isdir()]
+        return [p for p in self.listdir(*args, **kwargs) if p.isdir()]
 
-    def files(self, pattern=None):
+    def files(self, *args, **kwargs):
         """ D.files() -> List of the files in this directory.
 
         The elements of the list are Path objects.
         This does not walk into subdirectories (see :meth:`walkfiles`).
 
-        With the optional `pattern` argument, this only lists files
-        whose names match the given pattern.  For example,
-        ``d.files('*.pyc')``.
+        Accepts parameters to :meth:`listdir`.
         """
 
-        return [p for p in self.listdir(pattern) if p.isfile()]
+        return [p for p in self.listdir(*args, **kwargs) if p.isfile()]
 
-    def walk(self, pattern=None, errors='strict'):
+    def walk(self, match=None, errors='strict'):
         """ D.walk() -> iterator over files and subdirs, recursively.
 
         The iterator yields Path objects naming each child item of
@@ -564,6 +547,7 @@ class Path(text_type):
         reports the error via :func:`warnings.warn()`), and ``'ignore'``.
         `errors` may also be an arbitrary callable taking a msg parameter.
         """
+
         class Handlers:
             def strict(msg):
                 raise
@@ -578,6 +562,8 @@ class Path(text_type):
             raise ValueError("invalid errors parameter")
         errors = vars(Handlers).get(errors, errors)
 
+        match = matchers.load(match)
+
         try:
             childList = self.listdir()
         except Exception:
@@ -588,7 +574,7 @@ class Path(text_type):
             return
 
         for child in childList:
-            if pattern is None or child.fnmatch(pattern):
+            if match(child):
                 yield child
             try:
                 isdir = child.isdir()
@@ -600,92 +586,18 @@ class Path(text_type):
                 isdir = False
 
             if isdir:
-                for item in child.walk(pattern, errors):
+                for item in child.walk(errors=errors, match=match):
                     yield item
 
-    def walkdirs(self, pattern=None, errors='strict'):
+    def walkdirs(self, *args, **kwargs):
         """ D.walkdirs() -> iterator over subdirs, recursively.
-
-        With the optional `pattern` argument, this yields only
-        directories whose names match the given pattern.  For
-        example, ``mydir.walkdirs('*test')`` yields only directories
-        with names ending in ``'test'``.
-
-        The `errors=` keyword argument controls behavior when an
-        error occurs.  The default is ``'strict'``, which causes an
-        exception.  The other allowed values are ``'warn'`` (which
-        reports the error via :func:`warnings.warn()`), and ``'ignore'``.
         """
-        if errors not in ('strict', 'warn', 'ignore'):
-            raise ValueError("invalid errors parameter")
+        return (item for item in self.walk(*args, **kwargs) if item.isdir())
 
-        try:
-            dirs = self.dirs()
-        except Exception:
-            if errors == 'ignore':
-                return
-            elif errors == 'warn':
-                warnings.warn(
-                    "Unable to list directory '%s': %s"
-                    % (self, sys.exc_info()[1]),
-                    TreeWalkWarning)
-                return
-            else:
-                raise
-
-        for child in dirs:
-            if pattern is None or child.fnmatch(pattern):
-                yield child
-            for subsubdir in child.walkdirs(pattern, errors):
-                yield subsubdir
-
-    def walkfiles(self, pattern=None, errors='strict'):
+    def walkfiles(self, *args, **kwargs):
         """ D.walkfiles() -> iterator over files in D, recursively.
-
-        The optional argument `pattern` limits the results to files
-        with names that match the pattern.  For example,
-        ``mydir.walkfiles('*.tmp')`` yields only files with the ``.tmp``
-        extension.
         """
-        if errors not in ('strict', 'warn', 'ignore'):
-            raise ValueError("invalid errors parameter")
-
-        try:
-            childList = self.listdir()
-        except Exception:
-            if errors == 'ignore':
-                return
-            elif errors == 'warn':
-                warnings.warn(
-                    "Unable to list directory '%s': %s"
-                    % (self, sys.exc_info()[1]),
-                    TreeWalkWarning)
-                return
-            else:
-                raise
-
-        for child in childList:
-            try:
-                isfile = child.isfile()
-                isdir = not isfile and child.isdir()
-            except Exception:
-                if errors == 'ignore':
-                    continue
-                elif errors == 'warn':
-                    warnings.warn(
-                        "Unable to access '%s': %s"
-                        % (self, sys.exc_info()[1]),
-                        TreeWalkWarning)
-                    continue
-                else:
-                    raise
-
-            if isfile:
-                if pattern is None or child.fnmatch(pattern):
-                    yield child
-            elif isdir:
-                for f in child.walkfiles(pattern, errors):
-                    yield f
+        return (item for item in self.walk(*args, **kwargs) if item.isfile())
 
     def fnmatch(self, pattern, normcase=None):
         """ Return ``True`` if `self.name` matches the given `pattern`.
@@ -715,9 +627,31 @@ class Path(text_type):
         of all the files users have in their :file:`bin` directories.
 
         .. seealso:: :func:`glob.glob`
+
+        .. note:: Glob is **not** recursive, even when using ``**``.
+                  To do recursive globbing see :func:`walk`,
+                  :func:`walkdirs` or :func:`walkfiles`.
         """
         cls = self._next_class
         return [cls(s) for s in glob.glob(self / pattern)]
+
+    def iglob(self, pattern):
+        """ Return an iterator of Path objects that match the pattern.
+
+        `pattern` - a path relative to this directory, with wildcards.
+
+        For example, ``Path('/users').iglob('*/bin/*')`` returns an
+        iterator of all the files users have in their :file:`bin`
+        directories.
+
+        .. seealso:: :func:`glob.iglob`
+
+        .. note:: Glob is **not** recursive, even when using ``**``.
+                  To do recursive globbing see :func:`walk`,
+                  :func:`walkdirs` or :func:`walkfiles`.
+        """
+        cls = self._next_class
+        return (cls(s) for s in glob.iglob(self / pattern))
 
     #
     # --- Reading or writing an entire file at once.
@@ -728,8 +662,7 @@ class Path(text_type):
         Keyword arguments work as in :func:`io.open`.  If the file cannot be
         opened, an :class:`~exceptions.OSError` is raised.
         """
-        with io_error_compat():
-            return io.open(self, *args, **kwargs)
+        return io.open(self, *args, **kwargs)
 
     def bytes(self):
         """ Open this file, read all bytes, return them as a string. """
@@ -745,7 +678,7 @@ class Path(text_type):
            :example:
 
                >>> hash = hashlib.md5()
-               >>> for chunk in Path("path.py").chunks(8192, mode='rb'):
+               >>> for chunk in Path("CHANGES.rst").chunks(8192, mode='rb'):
                ...     hash.update(chunk)
 
             This will read the file by chunks of 8192 bytes.
@@ -767,19 +700,32 @@ class Path(text_type):
         with self.open(mode) as f:
             f.write(bytes)
 
-    def text(self, encoding=None, errors='strict'):
+    def read_text(self, encoding=None, errors=None):
         r""" Open this file, read it in, return the content as a string.
 
-        All newline sequences are converted to ``'\n'``.  Keyword arguments
-        will be passed to :meth:`open`.
+        Optional parameters are passed to :meth:`open`.
 
         .. seealso:: :meth:`lines`
         """
-        with self.open(mode='r', encoding=encoding, errors=errors) as f:
-            return U_NEWLINE.sub('\n', f.read())
+        with self.open(encoding=encoding, errors=errors) as f:
+            return f.read()
 
-    def write_text(self, text, encoding=None, errors='strict',
-                   linesep=os.linesep, append=False):
+    def read_bytes(self):
+        r"""Return the contents of this file as bytes."""
+        with self.open(mode='rb') as f:
+            return f.read()
+
+    def text(self, encoding=None, errors='strict'):
+        """Legacy function to read text.
+
+        Converts all newline sequences to ``\n``.
+        """
+        warnings.warn(".text is deprecated; use read_text", DeprecationWarning)
+        return U_NEWLINE.sub('\n', self.read_text(encoding, errors))
+
+    def write_text(
+        self, text, encoding=None, errors='strict', linesep=os.linesep, append=False
+    ):
         r""" Write the given text to this file.
 
         The default behavior is to overwrite any existing file;
@@ -843,7 +789,7 @@ class Path(text_type):
         conversion.
 
         """
-        if isinstance(text, text_type):
+        if isinstance(text, str):
             if linesep is not None:
                 text = U_NEWLINE.sub(linesep, text)
             text = text.encode(encoding or sys.getdefaultencoding(), errors)
@@ -867,18 +813,14 @@ class Path(text_type):
                 translated to ``'\n'``.  If ``False``, newline characters are
                 stripped off.  Default is ``True``.
 
-        This uses ``'U'`` mode.
-
         .. seealso:: :meth:`text`
         """
-        if encoding is None and retain:
-            with self.open('U') as f:
-                return f.readlines()
-        else:
-            return self.text(encoding, errors).splitlines(retain)
+        text = U_NEWLINE.sub('\n', self.read_text(encoding, errors))
+        return text.splitlines(retain)
 
-    def write_lines(self, lines, encoding=None, errors='strict',
-                    linesep=os.linesep, append=False):
+    def write_lines(
+        self, lines, encoding=None, errors='strict', linesep=os.linesep, append=False
+    ):
         r""" Write the given lines of text to this file.
 
         By default this overwrites any existing file at this path.
@@ -917,13 +859,12 @@ class Path(text_type):
         """
         with self.open('ab' if append else 'wb') as f:
             for line in lines:
-                isUnicode = isinstance(line, text_type)
+                isUnicode = isinstance(line, str)
                 if linesep is not None:
                     pattern = U_NL_END if isUnicode else NL_END
                     line = pattern.sub('', line) + linesep
                 if isUnicode:
-                    line = line.encode(
-                        encoding or sys.getdefaultencoding(), errors)
+                    line = line.encode(encoding or sys.getdefaultencoding(), errors)
                 f.write(line)
 
     def read_md5(self):
@@ -1007,46 +948,59 @@ class Path(text_type):
         return self.module.getatime(self)
 
     atime = property(
-        getatime, None, None,
+        getatime,
+        None,
+        None,
         """ Last access time of the file.
 
         .. seealso:: :meth:`getatime`, :func:`os.path.getatime`
-        """)
+        """,
+    )
 
     def getmtime(self):
         """ .. seealso:: :attr:`mtime`, :func:`os.path.getmtime` """
         return self.module.getmtime(self)
 
     mtime = property(
-        getmtime, None, None,
+        getmtime,
+        None,
+        None,
         """ Last-modified time of the file.
 
         .. seealso:: :meth:`getmtime`, :func:`os.path.getmtime`
-        """)
+        """,
+    )
 
     def getctime(self):
         """ .. seealso:: :attr:`ctime`, :func:`os.path.getctime` """
         return self.module.getctime(self)
 
     ctime = property(
-        getctime, None, None,
+        getctime,
+        None,
+        None,
         """ Creation time of the file.
 
         .. seealso:: :meth:`getctime`, :func:`os.path.getctime`
-        """)
+        """,
+    )
 
     def getsize(self):
         """ .. seealso:: :attr:`size`, :func:`os.path.getsize` """
         return self.module.getsize(self)
 
     size = property(
-        getsize, None, None,
+        getsize,
+        None,
+        None,
         """ Size of the file, in bytes.
 
         .. seealso:: :meth:`getsize`, :func:`os.path.getsize`
-        """)
+        """,
+    )
 
     if hasattr(os, 'access'):
+
         def access(self, mode):
             """ Return ``True`` if current user has access to this path.
 
@@ -1081,7 +1035,8 @@ class Path(text_type):
         .. seealso:: :attr:`owner`
         """
         desc = win32security.GetFileSecurity(
-            self, win32security.OWNER_SECURITY_INFORMATION)
+            self, win32security.OWNER_SECURITY_INFORMATION
+        )
         sid = desc.GetSecurityDescriptorOwner()
         account, domain, typecode = win32security.LookupAccountSid(None, sid)
         return domain + '\\' + account
@@ -1107,12 +1062,16 @@ class Path(text_type):
         get_owner = __get_owner_not_implemented
 
     owner = property(
-        get_owner, None, None,
+        get_owner,
+        None,
+        None,
         """ Name of the owner of this file or directory.
 
-        .. seealso:: :meth:`get_owner`""")
+        .. seealso:: :meth:`get_owner`""",
+    )
 
     if hasattr(os, 'statvfs'):
+
         def statvfs(self):
             """ Perform a ``statvfs()`` system call on this path.
 
@@ -1121,6 +1080,7 @@ class Path(text_type):
             return os.statvfs(self)
 
     if hasattr(os, 'pathconf'):
+
         def pathconf(self, name):
             """ .. seealso:: :func:`os.pathconf` """
             return os.pathconf(self, name)
@@ -1143,7 +1103,7 @@ class Path(text_type):
 
         .. seealso:: :func:`os.chmod`
         """
-        if isinstance(mode, string_types):
+        if isinstance(mode, str):
             mask = _multi_permission_mask(mode)
             mode = mask(self.stat().st_mode)
         os.chmod(self, mode)
@@ -1156,9 +1116,9 @@ class Path(text_type):
         .. seealso:: :func:`os.chown`
         """
         if hasattr(os, 'chown'):
-            if 'pwd' in globals() and isinstance(uid, string_types):
+            if 'pwd' in globals() and isinstance(uid, str):
                 uid = pwd.getpwnam(uid).pw_uid
-            if 'grp' in globals() and isinstance(gid, string_types):
+            if 'grp' in globals() and isinstance(gid, str):
                 gid = grp.getgrnam(gid).gr_gid
             os.chown(self, uid, gid)
         else:
@@ -1187,12 +1147,8 @@ class Path(text_type):
     def mkdir_p(self, mode=0o777):
         """ Like :meth:`mkdir`, but does not raise an exception if the
         directory already exists. """
-        try:
+        with contextlib.suppress(FileExistsError):
             self.mkdir(mode)
-        except OSError:
-            _, e, _ = sys.exc_info()
-            if e.errno != errno.EEXIST:
-                raise
         return self
 
     def makedirs(self, mode=0o777):
@@ -1203,12 +1159,8 @@ class Path(text_type):
     def makedirs_p(self, mode=0o777):
         """ Like :meth:`makedirs`, but does not raise an exception if the
         directory already exists. """
-        try:
+        with contextlib.suppress(FileExistsError):
             self.makedirs(mode)
-        except OSError:
-            _, e, _ = sys.exc_info()
-            if e.errno != errno.EEXIST:
-                raise
         return self
 
     def rmdir(self):
@@ -1219,13 +1171,10 @@ class Path(text_type):
     def rmdir_p(self):
         """ Like :meth:`rmdir`, but does not raise an exception if the
         directory is not empty or does not exist. """
-        try:
-            self.rmdir()
-        except OSError:
-            _, e, _ = sys.exc_info()
-            bypass_codes = errno.ENOTEMPTY, errno.EEXIST, errno.ENOENT
-            if e.errno not in bypass_codes:
-                raise
+        suppressed = FileNotFoundError, FileExistsError, DirectoryNotEmpty
+        with contextlib.suppress(suppressed):
+            with DirectoryNotEmpty.translate():
+                self.rmdir()
         return self
 
     def removedirs(self):
@@ -1236,12 +1185,9 @@ class Path(text_type):
     def removedirs_p(self):
         """ Like :meth:`removedirs`, but does not raise an exception if the
         directory is not empty or does not exist. """
-        try:
-            self.removedirs()
-        except OSError:
-            _, e, _ = sys.exc_info()
-            if e.errno != errno.ENOTEMPTY and e.errno != errno.EEXIST:
-                raise
+        with contextlib.suppress(FileExistsError, DirectoryNotEmpty):
+            with DirectoryNotEmpty.translate():
+                self.removedirs()
         return self
 
     # --- Modifying operations on files
@@ -1263,12 +1209,8 @@ class Path(text_type):
     def remove_p(self):
         """ Like :meth:`remove`, but does not raise an exception if the
         file does not exist. """
-        try:
+        with contextlib.suppress(FileNotFoundError):
             self.unlink()
-        except OSError:
-            _, e, _ = sys.exc_info()
-            if e.errno != errno.ENOENT:
-                raise
         return self
 
     def unlink(self):
@@ -1285,6 +1227,7 @@ class Path(text_type):
     # --- Links
 
     if hasattr(os, 'link'):
+
         def link(self, newpath):
             """ Create a hard link at `newpath`, pointing to this file.
 
@@ -1294,6 +1237,7 @@ class Path(text_type):
             return self._next_class(newpath)
 
     if hasattr(os, 'symlink'):
+
         def symlink(self, newlink=None):
             """ Create a symbolic link at `newlink`, pointing here.
 
@@ -1308,6 +1252,7 @@ class Path(text_type):
             return self._next_class(newlink)
 
     if hasattr(os, 'readlink'):
+
         def readlink(self):
             """ Return the path to which this symbolic link points.
 
@@ -1347,12 +1292,8 @@ class Path(text_type):
     def rmtree_p(self):
         """ Like :meth:`rmtree`, but does not raise an exception if the
         directory does not exist. """
-        try:
+        with contextlib.suppress(FileNotFoundError):
             self.rmtree()
-        except OSError:
-            _, e, _ = sys.exc_info()
-            if e.errno != errno.ENOENT:
-                raise
         return self
 
     def chdir(self):
@@ -1361,44 +1302,61 @@ class Path(text_type):
 
     cd = chdir
 
-    def merge_tree(self, dst, symlinks=False, *args, **kwargs):
+    def merge_tree(
+        self,
+        dst,
+        symlinks=False,
+        *,
+        copy_function=shutil.copy2,
+        ignore=lambda dir, contents: []
+    ):
         """
         Copy entire contents of self to dst, overwriting existing
         contents in dst with those in self.
 
-        If the additional keyword `update` is True, each
-        `src` will only be copied if `dst` does not exist,
-        or `src` is newer than `dst`.
+        Pass ``symlinks=True`` to copy symbolic links as links.
 
-        Note that the technique employed stages the files in a temporary
-        directory first, so this function is not suitable for merging
-        trees with large files, especially if the temporary directory
-        is not capable of storing a copy of the entire source tree.
+        Accepts a ``copy_function``, similar to copytree.
+
+        To avoid overwriting newer files, supply a copy function
+        wrapped in ``only_newer``. For example::
+
+            src.merge_tree(dst, copy_function=only_newer(shutil.copy2))
         """
-        update = kwargs.pop('update', False)
-        with tempdir() as _temp_dir:
-            # first copy the tree to a stage directory to support
-            #  the parameters and behavior of copytree.
-            stage = _temp_dir / str(hash(self))
-            self.copytree(stage, symlinks, *args, **kwargs)
-            # now copy everything from the stage directory using
-            #  the semantics of dir_util.copy_tree
-            distutils.dir_util.copy_tree(
-                stage,
-                dst,
-                preserve_symlinks=symlinks,
-                update=update,
-            )
+        dst = self._next_class(dst)
+        dst.makedirs_p()
+
+        sources = self.listdir()
+        _ignored = ignore(self, [item.name for item in sources])
+
+        def ignored(item):
+            return item.name in _ignored
+
+        for source in itertools.filterfalse(ignored, sources):
+            dest = dst / source.name
+            if symlinks and source.islink():
+                target = source.readlink()
+                target.symlink(dest)
+            elif source.isdir():
+                source.merge_tree(
+                    dest, symlinks=symlinks, copy_function=copy_function, ignore=ignore,
+                )
+            else:
+                copy_function(source, dest)
+
+        self.copystat(dst)
 
     #
     # --- Special stuff from os
 
     if hasattr(os, 'chroot'):
+
         def chroot(self):
             """ .. seealso:: :func:`os.chroot` """
             os.chroot(self)
 
     if hasattr(os, 'startfile'):
+
         def startfile(self):
             """ .. seealso:: :func:`os.startfile` """
             os.startfile(self)
@@ -1408,8 +1366,13 @@ class Path(text_type):
     # http://www.zopatista.com/python/2013/11/26/inplace-file-rewriting/
     @contextlib.contextmanager
     def in_place(
-            self, mode='r', buffering=-1, encoding=None, errors=None,
-            newline=None, backup_extension=None,
+        self,
+        mode='r',
+        buffering=-1,
+        encoding=None,
+        errors=None,
+        newline=None,
+        backup_extension=None,
     ):
         """
         A context in which a file may be re-written in-place with
@@ -1450,15 +1413,22 @@ class Path(text_type):
             pass
         os.rename(self, backup_fn)
         readable = io.open(
-            backup_fn, mode, buffering=buffering,
-            encoding=encoding, errors=errors, newline=newline,
+            backup_fn,
+            mode,
+            buffering=buffering,
+            encoding=encoding,
+            errors=errors,
+            newline=newline,
         )
         try:
             perm = os.fstat(readable.fileno()).st_mode
         except OSError:
             writable = open(
-                self, 'w' + mode.replace('r', ''),
-                buffering=buffering, encoding=encoding, errors=errors,
+                self,
+                'w' + mode.replace('r', ''),
+                buffering=buffering,
+                encoding=encoding,
+                errors=errors,
                 newline=newline,
             )
         else:
@@ -1467,8 +1437,11 @@ class Path(text_type):
                 os_mode |= os.O_BINARY
             fd = os.open(self, os_mode, perm)
             writable = io.open(
-                fd, "w" + mode.replace('r', ''),
-                buffering=buffering, encoding=encoding, errors=errors,
+                fd,
+                "w" + mode.replace('r', ''),
+                buffering=buffering,
+                encoding=encoding,
+                errors=errors,
                 newline=newline,
             )
             try:
@@ -1523,6 +1496,34 @@ class Path(text_type):
         return functools.partial(SpecialResolver, cls)
 
 
+class DirectoryNotEmpty(OSError):
+    @staticmethod
+    @contextlib.contextmanager
+    def translate():
+        try:
+            yield
+        except OSError as exc:
+            if exc.errno == errno.ENOTEMPTY:
+                raise DirectoryNotEmpty(*exc.args) from exc
+            raise
+
+
+def only_newer(copy_func):
+    """
+    Wrap a copy function (like shutil.copy2) to return
+    the dst if it's newer than the source.
+    """
+
+    @functools.wraps(copy_func)
+    def wrapper(src, dst, *args, **kwargs):
+        is_newer_dst = dst.exists() and dst.getmtime() >= src.getmtime()
+        if is_newer_dst:
+            return dst
+        return copy_func(src, dst, *args, **kwargs)
+
+    return wrapper
+
+
 class SpecialResolver(object):
     class ResolverScope:
         def __init__(self, paths, scope):
@@ -1540,8 +1541,7 @@ class SpecialResolver(object):
         not args and kwargs.setdefault('appname', None)
 
         vars(self).update(
-            path_class=path_class,
-            wrapper=appdirs.AppDirs(*args, **kwargs),
+            path_class=path_class, wrapper=appdirs.AppDirs(*args, **kwargs)
         )
 
     def __getattr__(self, scope):
@@ -1562,11 +1562,10 @@ class Multi:
     """
     A mix-in for a Path which may contain multiple Path separated by pathsep.
     """
+
     @classmethod
     def for_class(cls, path_cls):
         name = 'Multi' + path_cls.__name__
-        if PY2:
-            name = str(name)
         return type(name, (cls, path_cls), {})
 
     @classmethod
@@ -1584,22 +1583,18 @@ class Multi:
         """
         Multi-subclasses should use the parent class
         """
-        return next(
-            class_
-            for class_ in cls.__mro__
-            if not issubclass(class_, Multi)
-        )
+        return next(class_ for class_ in cls.__mro__ if not issubclass(class_, Multi))
 
 
-class tempdir(Path):
+class TempDir(Path):
     """
     A temporary directory via :func:`tempfile.mkdtemp`, and
     constructed with the same parameters that you can use
     as a context manager.
 
-    Example:
+    Example::
 
-        with tempdir() as d:
+        with TempDir() as d:
             # do stuff with the Path object "d"
 
         # here the directory is deleted automatically
@@ -1614,17 +1609,25 @@ class tempdir(Path):
 
     def __new__(cls, *args, **kwargs):
         dirname = tempfile.mkdtemp(*args, **kwargs)
-        return super(tempdir, cls).__new__(cls, dirname)
+        return super(TempDir, cls).__new__(cls, dirname)
 
     def __init__(self, *args, **kwargs):
         pass
 
     def __enter__(self):
-        return self
+        # TempDir should return a Path version of itself and not itself
+        # so that a second context manager does not create a second
+        # temporary directory, but rather changes CWD to the location
+        # of the temporary directory.
+        return self._next_class(self)
 
     def __exit__(self, exc_type, exc_value, traceback):
         if not exc_value:
             self.rmtree()
+
+
+# For backwards compatibility.
+tempdir = TempDir
 
 
 def _multi_permission_mask(mode):
@@ -1634,8 +1637,10 @@ def _multi_permission_mask(mode):
     >>> _multi_permission_mask('a=r,u+w')(0) == 0o644
     True
     """
+
     def compose(f, g):
         return lambda *args, **kwargs: g(f(*args, **kwargs))
+
     return functools.reduce(compose, map(_permission_mask, mode.split(',')))
 
 
@@ -1699,207 +1704,3 @@ def _permission_mask(mode):
         '=': lambda mask, target: target & retain ^ mask,
     }
     return functools.partial(op_map[op], mask)
-
-
-class CaseInsensitivePattern(text_type):
-    """
-    A string with a ``'normcase'`` property, suitable for passing to
-    :meth:`listdir`, :meth:`dirs`, :meth:`files`, :meth:`walk`,
-    :meth:`walkdirs`, or :meth:`walkfiles` to match case-insensitive.
-
-    For example, to get all files ending in .py, .Py, .pY, or .PY in the
-    current directory::
-
-        from path import Path, CaseInsensitivePattern as ci
-        Path('.').files(ci('*.py'))
-    """
-
-    @property
-    def normcase(self):
-        return __import__('ntpath').normcase
-
-
-class FastPath(Path):
-    """
-    Performance optimized version of Path for use
-    on embedded platforms and other systems with limited
-    CPU. See #115 and #116 for background.
-    """
-
-    def listdir(self, pattern=None):
-        children = os.listdir(self)
-        if pattern is None:
-            return [self / child for child in children]
-
-        pattern, normcase = self.__prepare(pattern)
-        return [
-            self / child
-            for child in children
-            if self._next_class(child).__fnmatch(pattern, normcase)
-        ]
-
-    def walk(self, pattern=None, errors='strict'):
-        class Handlers:
-            def strict(msg):
-                raise
-
-            def warn(msg):
-                warnings.warn(msg, TreeWalkWarning)
-
-            def ignore(msg):
-                pass
-
-        if not callable(errors) and errors not in vars(Handlers):
-            raise ValueError("invalid errors parameter")
-        errors = vars(Handlers).get(errors, errors)
-
-        if pattern:
-            pattern, normcase = self.__prepare(pattern)
-        else:
-            normcase = None
-
-        return self.__walk(pattern, normcase, errors)
-
-    def __walk(self, pattern, normcase, errors):
-        """ Prepared version of walk """
-        try:
-            childList = self.listdir()
-        except Exception:
-            exc = sys.exc_info()[1]
-            tmpl = "Unable to list directory '%(self)s': %(exc)s"
-            msg = tmpl % locals()
-            errors(msg)
-            return
-
-        for child in childList:
-            if pattern is None or child.__fnmatch(pattern, normcase):
-                yield child
-            try:
-                isdir = child.isdir()
-            except Exception:
-                exc = sys.exc_info()[1]
-                tmpl = "Unable to access '%(child)s': %(exc)s"
-                msg = tmpl % locals()
-                errors(msg)
-                isdir = False
-
-            if isdir:
-                for item in child.__walk(pattern, normcase, errors):
-                    yield item
-
-    def walkdirs(self, pattern=None, errors='strict'):
-        if errors not in ('strict', 'warn', 'ignore'):
-            raise ValueError("invalid errors parameter")
-
-        if pattern:
-            pattern, normcase = self.__prepare(pattern)
-        else:
-            normcase = None
-
-        return self.__walkdirs(pattern, normcase, errors)
-
-    def __walkdirs(self, pattern, normcase, errors):
-        """ Prepared version of walkdirs """
-        try:
-            dirs = self.dirs()
-        except Exception:
-            if errors == 'ignore':
-                return
-            elif errors == 'warn':
-                warnings.warn(
-                    "Unable to list directory '%s': %s"
-                    % (self, sys.exc_info()[1]),
-                    TreeWalkWarning)
-                return
-            else:
-                raise
-
-        for child in dirs:
-            if pattern is None or child.__fnmatch(pattern, normcase):
-                yield child
-            for subsubdir in child.__walkdirs(pattern, normcase, errors):
-                yield subsubdir
-
-    def walkfiles(self, pattern=None, errors='strict'):
-        if errors not in ('strict', 'warn', 'ignore'):
-            raise ValueError("invalid errors parameter")
-
-        if pattern:
-            pattern, normcase = self.__prepare(pattern)
-        else:
-            normcase = None
-
-        return self.__walkfiles(pattern, normcase, errors)
-
-    def __walkfiles(self, pattern, normcase, errors):
-        """ Prepared version of walkfiles """
-        try:
-            childList = self.listdir()
-        except Exception:
-            if errors == 'ignore':
-                return
-            elif errors == 'warn':
-                warnings.warn(
-                    "Unable to list directory '%s': %s"
-                    % (self, sys.exc_info()[1]),
-                    TreeWalkWarning)
-                return
-            else:
-                raise
-
-        for child in childList:
-            try:
-                isfile = child.isfile()
-                isdir = not isfile and child.isdir()
-            except Exception:
-                if errors == 'ignore':
-                    continue
-                elif errors == 'warn':
-                    warnings.warn(
-                        "Unable to access '%s': %s"
-                        % (self, sys.exc_info()[1]),
-                        TreeWalkWarning)
-                    continue
-                else:
-                    raise
-
-            if isfile:
-                if pattern is None or child.__fnmatch(pattern, normcase):
-                    yield child
-            elif isdir:
-                for f in child.__walkfiles(pattern, normcase, errors):
-                    yield f
-
-    def __fnmatch(self, pattern, normcase):
-        """ Return ``True`` if `self.name` matches the given `pattern`,
-        prepared version.
-        `pattern` - A filename pattern with wildcards,
-            for example ``'*.py'``. The pattern is expected to be normcase'd
-            already.
-        `normcase` - A function used to normalize the pattern and
-            filename before matching.
-        .. seealso:: :func:`Path.fnmatch`
-        """
-        return fnmatch.fnmatchcase(normcase(self.name), pattern)
-
-    def __prepare(self, pattern, normcase=None):
-        """ Prepares a fmatch_pattern for use with ``FastPath.__fnmatch`.
-        `pattern` - A filename pattern with wildcards,
-            for example ``'*.py'``. If the pattern contains a `normcase`
-            attribute, it is applied to the name and path prior to comparison.
-        `normcase` - (optional) A function used to normalize the pattern and
-            filename before matching. Defaults to :meth:`self.module`,
-            which defaults to :meth:`os.path.normcase`.
-        .. seealso:: :func:`FastPath.__fnmatch`
-        """
-        if not normcase:
-            normcase = getattr(pattern, 'normcase', self.module.normcase)
-        pattern = normcase(pattern)
-        return pattern, normcase
-
-    def fnmatch(self, pattern, normcase=None):
-        if not pattern:
-            raise ValueError("No pattern provided")
-
-        pattern, normcase = self.__prepare(pattern, normcase)
-        return self.__fnmatch(pattern, normcase)